moved open project to OpenProjectScenario
