--- src/project/internal/iopensaveprojectscenario.h
@@ -1,39 +1,0 @@
-/*
- * SPDX-License-Identifier: GPL-3.0-only
- * MuseScore-Studio-CLA-applies
- *
- * MuseScore Studio
- * Music Composition & Notation
- *
- * Copyright (C) 2022 MuseScore Limited and others
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License version 3 as
- * published by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-#pragma once
-
-#include "modularity/imoduleinterface.h"
-#include "types/retval.h"
-
-#include "inotationproject_fwd.h"
-#include "types/projecttypes.h"
-
-namespace mu::project {
-class IOpenSaveProjectScenario : MODULE_CONTEXT_INTERFACE
-{
-    INTERFACE_ID(IOpenSaveProjectScenario)
-
-public:
-    virtual void showCloudOpenError(const muse::Ret& ret) const = 0;
-};
-}
