+++ src/project/internal/projectactionscontroller.h
@@ -41,26 +41,14 @@
 #include "rcommand/icommanddispatcher.h"
 #include "multiwindows/imultiwindowsprovider.h"
 #include "multiwindows/iprojectprovider.h"
-#include "cloud/musescorecom/imusescorecomservice.h"
-#include "cloud/audiocom/iaudiocomservice.h"
 #include "print/iprintprovider.h"
-#include "iexportprojectscenario.h"
 #include "inotationreadersregister.h"
-#include "iopensaveprojectscenario.h"
+#include "iopenprojectscenario.h"
 #include "isaveprojectscenario.h"
-#include "imscmetareader.h"
-#include "io/ifilesystem.h"
-#include "notation/inotationconfiguration.h"
-#include "musesounds/imusesoundscheckupdatescenario.h"
-#include "musesounds/imusesamplercheckupdatescenario.h"
-#include "extensions/iextensionsprovider.h"
 
 #include "async/asyncable.h"
 
-#include "iprojectconfiguration.h"
-#include "iprojectcreator.h"
 #include "irecentfilescontroller.h"
-#include "iprojectautosaver.h"
 
 namespace mu::project {
 class ProjectActionsController : public IProjectCommandsController, public IProjectFilesController, public muse::mi::IProjectProvider,
@@ -69,23 +57,12 @@ class ProjectActionsController : public IProjectCommandsController, public IProj
     friend class ProjectActionsControllerTests;
 
 public:
-    muse::GlobalInject<IProjectConfiguration> configuration;
     muse::GlobalInject<muse::mi::IMultiWindowsProvider> multiwindowsProvider;
-    muse::GlobalInject<notation::INotationConfiguration> notationConfiguration;
-    muse::GlobalInject<muse::io::IFileSystem> fileSystem;
-    muse::GlobalInject<IMscMetaReader> mscMetaReader;
-    muse::GlobalInject<IProjectCreator> projectCreator;
-    muse::GlobalInject<muse::cloud::IMuseScoreComService> museScoreComService;
-    muse::GlobalInject<muse::cloud::IAudioComService> audioComService;
     muse::GlobalInject<INotationReadersRegister> readers;
     muse::GlobalInject<muse::IPlatformInteractive> platformInteractive;
-    muse::ContextInject<musesounds::IMuseSoundsCheckUpdateScenario> museSoundsCheckUpdateScenario = { this };
-    muse::ContextInject<musesounds::IMuseSamplerCheckUpdateScenario> museSamplerCheckUpdateScenario = { this };
     muse::ContextInject<IRecentFilesController> recentFilesController = { this };
-    muse::ContextInject<IProjectAutoSaver> projectAutoSaver = { this };
-    muse::ContextInject<IOpenSaveProjectScenario> openSaveProjectScenario = { this };
+    muse::ContextInject<IOpenProjectScenario> openProjectScenario = { this };
     muse::ContextInject<ISaveProjectScenario> saveProjectScenario = { this };
-    muse::ContextInject<IExportProjectScenario> exportProjectScenario = { this };
     muse::ContextInject<muse::actions::IActionsDispatcher> dispatcher = { this };
     muse::ContextInject<muse::rcommand::ICommandDispatcher> commandDispatcher = { this };
     muse::ContextInject<muse::IInteractive> interactive = { this };
@@ -138,18 +115,10 @@ private:
 
     muse::Ret openProject(const muse::rcommand::Params& params);
     muse::Ret openProject(const muse::io::path_t& path, const QString& displayNameOverride = QString());
-    void downloadAndOpenCloudProject(int scoreId, const QString& hash = QString(), const QString& secret = QString(), bool isOwner = true);
-    muse::Ret openMuseScoreUrl(const QUrl& url);
-    muse::Ret openScoreFromMuseScoreCom(const QUrl& url);
+
+    muse::Ret openPageIfNeed(muse::Uri pageUri);
 
     muse::Ret closeProject();
-
-    bool shouldRetryLoadAfterError(const muse::Ret& ret, const muse::io::path_t& filepath);
-    bool askIfUserAgreesToOpenProjectWithIncompatibleVersion(const std::string& errorText);
-    void warnFileTooNew(const muse::io::path_t& filepath);
-    bool askIfUserAgreesToOpenCorruptedProject(const muse::String& projectName, const std::string& errorText);
-    void warnProjectCriticallyCorrupted(const muse::String& projectName, const std::string& errorText);
-    void warnProjectCannotBeOpened(const muse::Ret& ret, const muse::io::path_t& filepath);
 
     muse::IInteractive::Button askAboutSavingScore(INotationProjectPtr project);
 
@@ -160,8 +129,6 @@ private:
 
     void revertCorruptedScoreToLastSaved();
 
-    RecentFile makeRecentFile(INotationProjectPtr project);
-
     muse::Ret importPdf();
     muse::Ret importAudioToScore();
 
@@ -170,18 +137,6 @@ private:
     muse::Ret continueLastSession();
 
     muse::Ret openProjectProperties();
-
-    muse::async::Promise<muse::io::path_t> selectScoreOpeningFile() const;
-
-    muse::RetVal<INotationProjectPtr> loadProject(const muse::io::path_t& filePath);
-    muse::Ret loadWithFallback(const std::shared_ptr<INotationProject>& project, const muse::io::path_t& loadPath,
-                               const std::string& format);
-    muse::Ret doOpenProject(const muse::io::path_t& filePath);
-    muse::Ret doOpenCloudProject(const muse::io::path_t& filePath, const CloudProjectInfo& info, bool isOwner = true);
-    muse::Ret doOpenCloudProjectOffline(const muse::io::path_t& filePath, const QString& displayNameOverride);
-
-    muse::Ret doFinishOpenProject();
-    muse::Ret openPageIfNeed(muse::Uri pageUri);
 
     muse::Ret exportScore();
     muse::Ret printScore();
@@ -194,7 +149,5 @@ private:
     muse::async::Notification m_needSaveChanged;
     muse::async::Notification m_hasSelectionChanged;
 
-    ProjectBeingDownloaded m_projectBeingDownloaded;
-    muse::async::Notification m_projectBeingDownloadedChanged;
 };
 }
