+++ src/project/tests/CMakeLists.txt
@@ -27,12 +27,12 @@ set(MODULE_TEST_SRC
     ${CMAKE_CURRENT_LIST_DIR}/mocks/projectconfigurationmock.h
     ${CMAKE_CURRENT_LIST_DIR}/mocks/recentfilescontrollermock.h
     ${CMAKE_CURRENT_LIST_DIR}/mocks/notationreadersregistermock.h
-    ${CMAKE_CURRENT_LIST_DIR}/mocks/opensaveprojectscenariomock.h
+    ${CMAKE_CURRENT_LIST_DIR}/mocks/openprojectscenariomock.h
     ${CMAKE_CURRENT_LIST_DIR}/mocks/projectautosavermock.h
     ${CMAKE_CURRENT_LIST_DIR}/mocks/projectcreatormock.h
     ${CMAKE_CURRENT_LIST_DIR}/mocks/saveprojectscenariomock.h
     ${CMAKE_CURRENT_LIST_DIR}/mscmetareadertests.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/projectactionscontrollertests.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/openprojectscenariotests.cpp
     ${CMAKE_CURRENT_LIST_DIR}/saveprojectscenariotests.cpp
     ${CMAKE_CURRENT_LIST_DIR}/templatesrepositorytest.cpp
 )
